Grokability master

+++ app/Http/Controllers/Components/ComponentsController.php
@@ -88,10 +88,16 @@ class ComponentsController extends Controller
 
         $component = $request->handleImages($component);
 
-        session()->put(['redirect_option' => $request->get('redirect_option')]);
+        if($request->get('redirect_option') === 'back'){
+            session()->put(['redirect_option' => 'index']);
+        } else {
+            session()->put(['redirect_option' => $request->get('redirect_option')]);
+        }
+
 
         if ($component->save()) {
-            return redirect()->to(Helper::getRedirectOption($request, $component->id, 'Components'))->with('success', trans('admin/components/message.create.success'));
+            return Helper::getRedirectOption($request, $component->id, 'Components')
+                ->with('success', trans('admin/components/message.create.success'));
         }
 
         return redirect()->back()->withInput()->withErrors($component->getErrors());
@@ -107,15 +113,14 @@ class ComponentsController extends Controller
      * @return \Illuminate\Contracts\View\View
      * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function edit($componentId = null)
+    public function edit(Component $component)
     {
-        if ($item = Component::find($componentId)) {
-            $this->authorize('update', $item);
 
-            return view('components/edit', compact('item'))->with('category_type', 'component');
-        }
-
-        return redirect()->route('components.index')->with('error', trans('admin/components/message.does_not_exist'));
+            $this->authorize('update', $component);
+            session()->put('back_url', url()->previous());
+            return view('components/edit')
+                ->with('item', $component)
+                ->with('category_type', 'component');
     }
 
 
@@ -130,11 +135,8 @@ class ComponentsController extends Controller
      * @throws \Illuminate\Auth\Access\AuthorizationException
      * @since [v3.0]
      */
-    public function update(ImageUploadRequest $request, $componentId = null)
+    public function update(ImageUploadRequest $request, Component $component)
     {
-        if (is_null($component = Component::find($componentId))) {
-            return redirect()->route('components.index')->with('error', trans('admin/components/message.does_not_exist'));
-        }
         $min = $component->numCheckedOut();
         $validator = Validator::make($request->all(), [
             'qty' => "required|numeric|min:$min",
@@ -169,7 +171,8 @@ class ComponentsController extends Controller
         session()->put(['redirect_option' => $request->get('redirect_option')]);
 
         if ($component->save()) {
-            return redirect()->to(Helper::getRedirectOption($request, $component->id, 'Components'))->with('success', trans('admin/components/message.update.success'));
+            return Helper::getRedirectOption($request, $component->id, 'Components')
+                ->with('success', trans('admin/components/message.update.success'));
         }
 
         return redirect()->back()->withInput()->withErrors($component->getErrors());
@@ -201,6 +204,10 @@ class ComponentsController extends Controller
             }
         }
 
+        if ($component->numCheckedOut() > 0) {
+            return redirect()->route('components.index')->with('error', trans('admin/components/message.delete.error_qty'));
+        }
+
         $component->delete();
 
         return redirect()->route('components.index')->with('success', trans('admin/components/message.delete.success'));
@@ -216,17 +223,9 @@ class ComponentsController extends Controller
      * @return \Illuminate\Contracts\View\View
      * @throws \Illuminate\Auth\Access\AuthorizationException
      */
-    public function show($componentId = null)
+    public function show(Component $component)
     {
-        $component = Component::find($componentId);
-
-        if (isset($component->id)) {
             $this->authorize('view', $component);
-
             return view('components/view', compact('component'));
-        }
-        // Redirect to the user management page
-        return redirect()->route('components.index')
-            ->with('error', trans('admin/components/message.does_not_exist'));
     }
 }
